requirements.txt
pymongo==4.6.1
psycopg2-binary==2.9.10
elasticsearch==8.14.0
